get_operation -> get_instruction and bitflags
it now returns a struct with more information instead of just the op.
this is done so it's simpler to add new mnemonics in the future.

also added the bitflags crate to get named bitmasks. could not be arsed
to figure out how to do it without casting to u8 so w/e dependency wahey