handling immediate to register operations seperately
this is because we were trying to cast the bits to a certain enum
each time which would cause a panic incase it fails to do so,
failing to do this is actually expected behavior because the next
branch will handle the edge cases which cause this to happen.

- made better error messages too btw.
- added expected_mnemonic to the get_instruction tests because we're
starting to add support for diff ones.
- some annoying vscode debug configurations.