about infinite loop and other stuff.
Infinite loop was supposed to happen since the simulation was not supposed to be done on that listing.

also, simulating is now a boolean flag since not all listings want to do it.

on top of this I started making a test for listing_0039 that also tests the simulation but that one does not yet work.