Fixed logic error in uninitialized registers. Also fixed a branch I did not handle with memory storing.
There was a logic error where if a register or memory location was initialized with 0, it would remain in an Uninitialized state but that is obviously not correct since it is initialized with 0.

The last instruction needs fixing and after that the 0039 test will pass.

On top of all this, I added branch I forgot to handle with the mov [bp + 0], ch instruction, I luckily spotted this with the debugger because memory was not getting initialized correctly.