Fixed problem with memory locations.
Before this fix, memory locations were not correctly aware of the previous value in the memory location. I fixed this and it was quite a dumb mistake because I was trying to get the value of the rm_register with the reg_register. Yikes af. Fixed thought so we gucci.