Fixed logic in instruction_has_immediate_value_in_rm_register.
The problem was that we checked for DirectMemoryOperation in the branch even though it does not matter at all when the instruction is ImmediateToRegisterMOV. This lead into the rm_immediate variable to not be populated and the format_instruction function returned mov bp, 0 when it should return mov bp, 6. This is now fixed and listing_0043 passes now.